Export data function
Export all the users games to a csv file and provides download.  I placed the button on the profile page not sure if that is gona be the permanent location.  But looks like it is working.  Im probably gona add more stats and can improve how the csv is formed but this atleast works for now.